test(daemon): update service load-state fixtures
